garmage_jigsaw/requirement.txt
# Basic
wheel
pip
tqdm
wandb==0.13.10

# Core PyTorch & Model
numpy==1.23.5
torch==1.10.2
torchvision==0.11.3
torch_geometric
pytorch-lightning==1.9.1
torch-cluster==1.6.0
cudatoolkit==11.8
diffusers==0.32.1

# 3D & geometry
open3d==0.11.2
trimesh==3.20.0
libigl==2.5.1

# ML & data
scipy==1.10.0
scikit-learn==1.2.0
sklearn
pandas==1.5.3
einops==0.6.0
easydict==1.9
matplotlib==3.7.0
opencv-python==4.12

